미션 오버플로우 fix

+++ Assets/Scripts/Stat/Model/StatsSO.cs
@@ -295,7 +295,7 @@ public class StatsSO : ScriptableObject
 
     public void AddGold(BigInteger gold)
     {
-        MissionManager.Instance.EarnedGold((int)gold);
+        MissionManager.Instance.EarnedGold(gold);
         Gold.Value += gold;
         DataManager.Instance.SaveData();
     }
